A5.1: Combine Art IV Sec 1 of the Con & Art III (Locals) Sec 1 of the Bylaws
Part of the Constitution's Article IV deals with Locals; this will be combined with Article III in the Bylaws, the entirety of which deals with Locals.

These two Sections in particular will be combined because they both discuss the process of chartering a Local. One sentence of the Constitution ("The NPC may charter Locals of 10 or more in special circumstances as defined in the Bylaws") is deleted because it is now redundant.

NOTE: With this deletion, does the first sentence have to be changed to make clear that sometimes a Local can be chartered with only 10-14 signatures? Consult with lawyers.