A6.1: Combine Art IV Sec 3 of the Con & Art IV Sec 1 of the Bylaws
This Article will be a combination of Article IV (State and Regional Organizations) in the Bylaws and the part of Article IV in the Constitution that deals with state and regional organizations.

These two Sections are combined because they both describe the process of chartering a state organization. One sentence from the Constitution ("Upon petition of two or more Locals, the NPC shall consider the application for State organizations") is deleted because it is now redundant. A few words were also changed in the second-to-last sentence due to it being moved into a new Bylaws.

NOTE: Does anything change because this new Section does not contain the word "petition"?