A6.5: Combine Art IV Sec 5 in the Con & Art IV Sec 5 in the Bylaws
These two Sections are combined because they both deal with revoking a charter. Note that one sentence from the Constitution ("Criteria for the revocation of charters are prescribed in the Bylaws") is deleted because it is now redundant. A grammatical change was also made in the first sentence, and the Article number in the last sentence was changed to reflect the new number in the updated Bylaws.